+++ samples/actionbarsherlock/src/com/manuelpeinado/fadingactionbar/samples/actionbarsherlock/HomeActivity.java
@@ -26,6 +26,8 @@ import android.widget.ArrayAdapter;
 import android.widget.ListView;
 
 import com.actionbarsherlock.app.SherlockListActivity;
+import com.manuelpeinado.fadingactionbar.samples.actionbarcompat.ActivityInfo;
+import com.manuelpeinado.fadingactionbar.samples.actionbarcompat.HeaderOverlayActivity;
 import com.manuelpeinado.fadingactionbar.samples.actionbarsherlock.R;
 
 public class HomeActivity extends SherlockListActivity {
@@ -36,7 +38,8 @@ public class HomeActivity extends SherlockListActivity {
             new ActivityInfo(LightActionBarActivity.class, R.string.activity_title_light_ab),
             new ActivityInfo(SampleFragmentActivity.class, R.string.activity_title_fragment),
             new ActivityInfo(NoParallaxActivity.class, R.string.activity_title_no_parallax),
-            new ActivityInfo(NavigationDrawerActivity.class, R.string.activity_title_navigation));
+            new ActivityInfo(NavigationDrawerActivity.class, R.string.activity_title_navigation),
+            new ActivityInfo(HeaderOverlayActivity.class, R.string.activity_title_header_overlay));
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {