+++ samples/stock/src/com/manuelpeinado/fadingactionbar/demo/HeaderOverlayActivity.java
@@ -1,0 +1,76 @@
+/*
+ * Copyright (C) 2013 Manuel Peinado
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.manuelpeinado.fadingactionbar.demo;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.util.ArrayList;
+
+import android.app.Activity;
+import android.os.Bundle;
+import android.view.Menu;
+import android.widget.ArrayAdapter;
+import android.widget.ListView;
+
+import com.manuelpeinado.fadingactionbar.FadingActionBarHelper;
+
+public class HeaderOverlayActivity extends Activity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        FadingActionBarHelper helper = new FadingActionBarHelper()
+            .actionBarBackground(R.drawable.ab_background)
+            .headerLayout(R.layout.header)
+            .contentLayout(R.layout.activity_listview)
+            .headerOverlayLayout(R.layout.header_overlay);
+        setContentView(helper.createView(this));
+        helper.initActionBar(this);
+
+        ListView listView = (ListView) findViewById(android.R.id.list);
+        ArrayList<String> items = loadItems(R.raw.nyc_sites);
+        ArrayAdapter<String> adapter = new ArrayAdapter<String>(this, android.R.layout.simple_list_item_1, items);
+        listView.setAdapter(adapter);
+    }
+
+    /**
+     * @return A list of Strings read from the specified resource
+     */
+    private ArrayList<String> loadItems(int rawResourceId) {
+        try {
+            ArrayList<String> countries = new ArrayList<String>();
+            InputStream inputStream = getResources().openRawResource(rawResourceId);
+            BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream));
+            String line;
+            while ((line = reader.readLine()) != null) {
+                countries.add(line);
+            }
+            reader.close();
+            return countries;
+        } catch (IOException e) {
+            return null;
+        }
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.activity_menu, menu);
+        return true;
+    }
+}