+++ samples/stock/src/com/manuelpeinado/fadingactionbar/demo/HomeActivity.java
@@ -34,7 +34,8 @@ public class HomeActivity extends ListActivity {
             new ActivityInfo(LightActionBarActivity.class, R.string.activity_title_light_ab),
             new ActivityInfo(SampleFragmentActivity.class, R.string.activity_title_fragment),
             new ActivityInfo(NoParallaxActivity.class, R.string.activity_title_no_parallax),
-            new ActivityInfo(NavigationDrawerActivity.class, R.string.activity_title_navigation));
+            new ActivityInfo(NavigationDrawerActivity.class, R.string.activity_title_navigation),
+            new ActivityInfo(HeaderOverlayActivity.class, R.string.activity_title_header_overlay));
     
     @Override
     protected void onCreate(Bundle savedInstanceState) {